requirements.txt
fastapi==0.95.0
uvicorn==0.23.0

sqlalchemy==2.0.21
psycopg2-binary==2.9.6

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-dotenv==1.0.0

ecdsa==0.19.0
pyasn1==0.6.0
rsa==4.9
six==1.16.0

python-multipart==0.0.6
